Better handling of edge cases, changed upscaling to exactly 2.
By interpolating the out of border pixels using edge values, the pixels near the edge can be upsampled using DCCI.

The interpolation proposed here is a 0th order interpolation. The out of border pixels are simply assumed to be equal to the edge pixel values.
A linear (1st order) or squared (2nd order) interpolation would probably be better, but I don't know if it is worth the computational complexity.